+++ src/reg_key.rs
@@ -166,10 +166,34 @@ impl RegKey {
         path: P,
         perms: Registry::REG_SAM_FLAGS,
     ) -> io::Result<RegKey> {
+        self.open_subkey_with_options_flags(path, 0, perms)
+    }
+
+    /// Open subkey with desired permissions and options.
+    /// Will open another handle to itself if `path` is an empty string.
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// # use std::error::Error;
+    /// # use winreg::RegKey;
+    /// # use winreg::enums::*;
+    /// # fn main() -> Result<(), Box<dyn Error>> {
+    /// let hklm = RegKey::predef(HKEY_LOCAL_MACHINE);
+    /// hklm.open_subkey_with_options_flags("SOFTWARE\\LinkKey", REG_OPTION_OPEN_LINK, KEY_READ)?;
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn open_subkey_with_options_flags<P: AsRef<OsStr>>(
+        &self,
+        path: P,
+        options: Registry::REG_OPEN_CREATE_OPTIONS,
+        perms: Registry::REG_SAM_FLAGS,
+    ) -> io::Result<RegKey> {
         let c_path = to_utf16(path);
         let mut new_hkey: HKEY = 0;
         match unsafe {
-            Registry::RegOpenKeyExW(self.hkey, c_path.as_ptr(), 0, perms, &mut new_hkey)
+            Registry::RegOpenKeyExW(self.hkey, c_path.as_ptr(), options, perms, &mut new_hkey)
         } {
             0 => Ok(RegKey { hkey: new_hkey }),
             err => werr!(err),
@@ -194,13 +218,25 @@ impl RegKey {
         t: &Transaction,
         perms: Registry::REG_SAM_FLAGS,
     ) -> io::Result<RegKey> {
+        self.open_subkey_transacted_with_options_flags(path, t, 0, perms)
+    }
+
+    /// Part of `transactions` feature.
+    #[cfg(feature = "transactions")]
+    pub fn open_subkey_transacted_with_options_flags<P: AsRef<OsStr>>(
+        &self,
+        path: P,
+        t: &Transaction,
+        options: Registry::REG_OPEN_CREATE_OPTIONS,
+        perms: Registry::REG_SAM_FLAGS,
+    ) -> io::Result<RegKey> {
         let c_path = to_utf16(path);
         let mut new_hkey: HKEY = 0;
         match unsafe {
             Registry::RegOpenKeyTransactedW(
                 self.hkey,
                 c_path.as_ptr(),
-                0,
+                options,
                 perms,
                 &mut new_hkey,
                 t.handle,
@@ -246,6 +282,15 @@ impl RegKey {
         path: P,
         perms: Registry::REG_SAM_FLAGS,
     ) -> io::Result<(RegKey, RegDisposition)> {
+        self.create_subkey_with_options_flags(path, REG_OPTION_NON_VOLATILE, perms)
+    }
+
+    pub fn create_subkey_with_options_flags<P: AsRef<OsStr>>(
+        &self,
+        path: P,
+        options: Registry::REG_OPEN_CREATE_OPTIONS,
+        perms: Registry::REG_SAM_FLAGS,
+    ) -> io::Result<(RegKey, RegDisposition)> {
         let c_path = to_utf16(path);
         let mut new_hkey: HKEY = 0;
         let mut disp_buf: u32 = 0;
@@ -255,7 +300,7 @@ impl RegKey {
                 c_path.as_ptr(),
                 0,
                 ptr::null_mut(),
-                Registry::REG_OPTION_NON_VOLATILE,
+                options,
                 perms,
                 ptr::null_mut(),
                 &mut new_hkey,
@@ -288,6 +333,18 @@ impl RegKey {
         t: &Transaction,
         perms: Registry::REG_SAM_FLAGS,
     ) -> io::Result<(RegKey, RegDisposition)> {
+        self.create_subkey_transacted_with_options_flags(path, t, REG_OPTION_NON_VOLATILE, perms)
+    }
+
+    /// Part of `transactions` feature.
+    #[cfg(feature = "transactions")]
+    pub fn create_subkey_transacted_with_options_flags<P: AsRef<OsStr>>(
+        &self,
+        path: P,
+        t: &Transaction,
+        options: Registry::REG_OPEN_CREATE_OPTIONS,
+        perms: Registry::REG_SAM_FLAGS,
+    ) -> io::Result<(RegKey, RegDisposition)> {
         let c_path = to_utf16(path);
         let mut new_hkey: HKEY = 0;
         let mut disp_buf: u32 = 0;
@@ -297,7 +354,7 @@ impl RegKey {
                 c_path.as_ptr(),
                 0,
                 ptr::null_mut(),
-                Registry::REG_OPTION_NON_VOLATILE,
+                options,
                 perms,
                 ptr::null_mut(),
                 &mut new_hkey,
@@ -608,7 +665,7 @@ impl RegKey {
             match unsafe {
                 Registry::RegQueryValueExW(
                     self.hkey,
-                    c_name.as_ptr() as *const u16,
+                    c_name.as_ptr(),
                     ptr::null_mut(),
                     &mut buf_type,
                     buf.as_mut_ptr() as *mut u8,
@@ -960,7 +1017,7 @@ pub struct EnumKeys<'key> {
     index: u32,
 }
 
-impl<'key> Iterator for EnumKeys<'key> {
+impl Iterator for EnumKeys<'_> {
     type Item = io::Result<String>;
 
     fn next(&mut self) -> Option<io::Result<String>> {
@@ -985,7 +1042,7 @@ pub struct EnumValues<'key> {
     index: u32,
 }
 
-impl<'key> Iterator for EnumValues<'key> {
+impl Iterator for EnumValues<'_> {
     type Item = io::Result<(String, RegValue)>;
 
     fn next(&mut self) -> Option<io::Result<(String, RegValue)>> {